# Improve precision of Quantity.AsApproximateFloat64