This improves the precision of Quantity.AsApproximateFloat64, by way of example:

`decQuantity(7*1024*1024, -1, BinarySI)`
Before: 734003.2000000001,
After: 734003.2

/kind cleanup
/sig api-machinery

```release-note
Fixed floating point error when recording and displaying resource quantities
```

@BoSunesen