@thockin 2023-08-15
Is there a bug or something that shows why this matters?  I'm not super familiar with thise code, and it is used in just 1 or 2 places.  Does this really matter?