@intUnderflow 2023-08-17
There was a bug we encountered as end users that made us stumble across this, @BoSunesen do you have context on what the bug was again?

Even without that though in my opinion this is justifiable for inclusion on its own as a quality of life improvement, especially as we've already written the code for it.