@intUnderflow 2024-01-08
@thockin This is causing us problems in our usage of Kubernetes because it means comparisons that should pass end up failing due to floating point errors:

> current resources are not equal to spec resources, "&Resources{CpuCores:0.7,MemoryInMb:2048,DiskInMb:2048,FileDescriptors:5000,GpuCores:0,DedicatedNodes:0,}" != "&Resources{CpuCores:0.7000000000000001,MemoryInMb:2048,DiskInMb:2048,FileDescriptors:5000,GpuCores:0,DedicatedNodes:0,}"