@thockin 2024-01-23
Looking at the change and the tests, this SEEMS semantically OK.  What I am worried about (if I understand it right) is that there used to be a "fast path" which bypassed `big.Int` and `big.Float` and now there isn't.  How much does that hurt?

It's not really commented as to WHY this was implemented the way it was.  It's clearly trading off accuracy for performance, in a function EXPLICITLY named to say "approximate".

There appear to be 2 consumers, other than tests:

pkg/scheduler/metrics/resources/resources.go

This looks to be in the path of collecting pod metrics, which...seems important?  I can't say for sure.  @smarterclayton touched it, not sure if he's right to assess this.

staging/src/k8s.io/apiserver/pkg/cel/library/quantity.go

@jpbetz tells me this CEL support is new, so it would not be a "regression"

Could I arm-twist you into running the bechmark for this (BenchmarkQuantityAsApproximateFloat64) before and after this PR and posting the results here?