@thockin 2024-07-13
Before:
```
$ go test ./staging/src/k8s.io/apimachinery/pkg/api/resource/ -bench=Approx
goos: linux
goarch: amd64
pkg: k8s.io/apimachinery/pkg/api/resource
cpu: Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz
BenchmarkQuantityAsApproximateFloat64-6   	93582822	        11.56 ns/op
PASS
ok  	k8s.io/apimachinery/pkg/api/resource	1.123s
```

After:
```
$ go test ./staging/src/k8s.io/apimachinery/pkg/api/resource/ -bench=Approx
goos: linux
goarch: amd64
pkg: k8s.io/apimachinery/pkg/api/resource
cpu: Intel(R) Xeon(R) W-2135 CPU @ 3.70GHz
BenchmarkQuantityAsApproximateFloat64-6   	 2890150	       420.2 ns/op
PASS
ok  	k8s.io/apimachinery/pkg/api/resource	1.661s
```

So, that particular call is now 35x slower.  Does that matter?  This is not so frequently called but I hate laying traps.

@alculquicondor for viz on scheduling.

@intUnderflow On the one hand, we all know that exact-comparisons with floats is trouble.  On the other hand, if we can make it better, that's nice.  What if you instead offered this as a NEW method, like `AsFloat64()` or `AsApproximateFloat64Slow()` with good comments explaining the precision/performance tradeoffs between the two?  (and a new benchmark func :)