@chenxinlong 2024-09-14
Same problem here using k8s.io/apimachinery v0.23.1
```go
q1 := resource.MustParse("9100m")
fmt.Println(q1.AsApproximateFloat64()) // 9.1

q2 := resource.MustParse("9200m")
fmt.Println(q2.AsApproximateFloat64()) // 9.200000000000001
```