@thockin 2024-09-14
I don't object to exposing this as a somewhat more accurate float64, but I really need to point out, again, that __ANY TIME__ you compare floating point values, there's a risk of this sort of thing.  You simply should not be relying on EXACT comparisons with floats.